Refactor: Rename "dao" layer to "repository"
