+++ src/main/java/com/sankdev/fluffybest/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package com.sankdev.fluffybest.dao;
+package com.sankdev.fluffybest.repository;
 
 import com.sankdev.fluffybest.entity.Product;
 import org.springframework.data.domain.Page;
