+++ src/main/java/com/sankdev/fluffybest/repository/location/CountryRepository.java
@@ -1,4 +1,4 @@
-package com.sankdev.fluffybest.dao.location;
+package com.sankdev.fluffybest.repository.location;
 
 import com.sankdev.fluffybest.entity.location.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
